break out MachineRequestSerializer
