+++ api/serializers/__init__.py
@@ -23,3 +23,4 @@ from .paginated_instance_serializer import PaginatedInstanceSerializer
 from .machine_export_serializer import MachineExportSerializer
 from .license_serializer import LicenseSerializer
 from .post_license_serializer import POST_LicenseSerializer
+from .machine_request_serializer import MachineRequestSerializer
