+++ api/serializers/machine_request_serializer.py
@@ -1,0 +1,48 @@
+from core.models.machine_request import MachineRequest
+from rest_framework import serializers
+from .new_threshold_field import NewThresholdField
+from .license_serializer import LicenseSerializer
+
+
+class MachineRequestSerializer(serializers.ModelSerializer):
+    """
+    """
+    instance = serializers.SlugRelatedField(slug_field='provider_alias')
+    status = serializers.CharField(default="pending")
+    parent_machine = serializers.SlugRelatedField(slug_field='identifier',
+                                                  read_only=True)
+
+    sys = serializers.CharField(default="", source='iplant_sys_files',
+                                required=False)
+    software = serializers.CharField(default="No software listed",
+                                     source='installed_software',
+                                     required=False)
+    exclude_files = serializers.CharField(default="", required=False)
+    shared_with = serializers.CharField(source="access_list", required=False)
+
+    name = serializers.CharField(source='new_machine_name')
+    provider = serializers.SlugRelatedField(
+        slug_field='uuid', source='new_machine_provider')
+    owner = serializers.SlugRelatedField(slug_field='username',
+                                         source='new_machine_owner')
+    vis = serializers.CharField(source='new_machine_visibility')
+    version = serializers.CharField(source='new_machine_version',
+            required=False)
+    fork = serializers.BooleanField(source='new_machine_forked',
+            required=False)
+    description = serializers.CharField(source='new_machine_description',
+                                        required=False)
+    tags = serializers.CharField(source='new_machine_tags', required=False)
+    threshold = NewThresholdField(source='new_machine_threshold')
+    #TODO: Convert to 'LicenseField' and allow POST of ID instead of
+    #      full-object. for additional support for the image creator
+    licenses = LicenseSerializer(source='new_machine_licenses.all', many=True)
+    new_machine = serializers.SlugRelatedField(slug_field='identifier',
+                                               required=False)
+
+    class Meta:
+        model = MachineRequest
+        fields = ('id', 'instance', 'status', 'name', 'owner', 'provider',
+                  'vis', 'description', 'tags', 'sys', 'software',
+                  'threshold', 'fork', 'version',
+                  'shared_with', 'licenses', 'new_machine')
